+++ src/main/java/com/sda/project/repository/PetRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PetRepository extends JpaRepository<Pet, Integer> {
+public interface PetRepository extends JpaRepository<Pet, Long> {
+
     Optional<Pet> findByNameIgnoreCase(String name);
 
     Optional<List<Pet>> findByCategory(String category);
-
 
     // TODO: step 3
